feat #17: устранение проблем с запуском и со сборкой контейнера
Update

- *.yml will use Unix (LF)
- провести тестирование и устранения проблем с конфигурациями при запуске и сборки контейнеров
- убрать из проекта Infrastructure для Mongo & InMemmory ("лишний код" которые пока не будут использоваться)
- сделать отдельный проект Infrastructure.Postgres (c 2 DBCOntext Main & Indentity & Test)
- добавить Readme по сборке контейнера и запуску в базовых конфигурациях
