+++ SeaWind.sln
@@ -14,11 +14,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{019FED03
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "src\backend\Application\Application.csproj", "{D6BA57F0-6F83-4DC8-AB4F-A5BFF65F9A6E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure", "src\backend\Infrastructure\Infrastructure.csproj", "{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.UnitTests", "tests\Api.UnitTests\Api.UnitTests.csproj", "{FFC8F8BA-CB00-4AB9-BA30-3C7137B16AE6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application.UnitTests", "tests\Application.UnitTests\Application.UnitTests.csproj", "{B6F7D088-C83D-46A1-BDC4-9D0218ACABDC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Infrastructure.Postgres", "src\backend\Infrastructure.Postgres\Infrastructure.Postgres.csproj", "{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,10 +34,6 @@ Global
 		{D6BA57F0-6F83-4DC8-AB4F-A5BFF65F9A6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D6BA57F0-6F83-4DC8-AB4F-A5BFF65F9A6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D6BA57F0-6F83-4DC8-AB4F-A5BFF65F9A6E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FFC8F8BA-CB00-4AB9-BA30-3C7137B16AE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FFC8F8BA-CB00-4AB9-BA30-3C7137B16AE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FFC8F8BA-CB00-4AB9-BA30-3C7137B16AE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -46,6 +42,10 @@ Global
 		{B6F7D088-C83D-46A1-BDC4-9D0218ACABDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B6F7D088-C83D-46A1-BDC4-9D0218ACABDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B6F7D088-C83D-46A1-BDC4-9D0218ACABDC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -53,9 +53,9 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{EAC707C0-BCEE-48AA-8C9E-89CEEFA26CCE} = {EAF2A4AC-4A44-4DE0-BECA-E72C68007453}
 		{D6BA57F0-6F83-4DC8-AB4F-A5BFF65F9A6E} = {EAF2A4AC-4A44-4DE0-BECA-E72C68007453}
-		{9F8662F8-B91D-4C2C-B2BE-C1D52B68298D} = {EAF2A4AC-4A44-4DE0-BECA-E72C68007453}
 		{FFC8F8BA-CB00-4AB9-BA30-3C7137B16AE6} = {019FED03-8ADF-4C25-8B4B-A2E69B26B1D8}
 		{B6F7D088-C83D-46A1-BDC4-9D0218ACABDC} = {019FED03-8ADF-4C25-8B4B-A2E69B26B1D8}
+		{0D9A3B91-6F59-4B22-8DC3-9FE4835A4456} = {EAF2A4AC-4A44-4DE0-BECA-E72C68007453}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {563DA8B7-95B5-4369-9D07-AE43505E9535}
