+++ docker/.env.example
@@ -1,21 +1,20 @@
-#!!!! No WEB parameters
-# API
-ASPNETCORE_ENVIRONMENT=Production
-#ASPNETCORE_ENVIRONMENT=Development
-ASPNETCORE_URLS=http://+:8080
-HTTP_PORT=5000
+ASPNETCORE_ENVIRONMENT=Development
+HTTP_PORT_EXT=5000
+HTTP_PORT_INT=8080
 
 # DB
-POSTGRES_HOST=db
 POSTGRES_DB=workshopcode
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
-POSTGRES_PORT=55432
+POSTGRES_PORT_EXT=55432
+POSTGRES_PORT_INT=5432
 
 # App DB user (ограниченный)
 APP_DB=workshopcode
 APP_USER=wc_app
 APP_PASSWORD=wc_app_pwd
 
-# EF Connection (перекрывает appsettings.Development.json)
-ConnectionStrings__Default=Host=db;Port=5432;Database=workshopcode;Username=wc_app;Password=wc_app_pwd;Pooling=true
+# Внутри контейнера используется своя собственная сеть 
+# в контейнере свой <ConnectionStrings__Default> (перекрывает appsettings.Development.json)
+# см. docker-compose.yml
+
