+++ docker/docker-compose.yml
@@ -4,7 +4,7 @@ services:
   db:
     image: postgres:16
     ports:
-      - "${POSTGRES_PORT}:5432"
+      - "${POSTGRES_PORT_EXT}:${POSTGRES_PORT_INT}"
     restart: unless-stopped
     environment:
       POSTGRES_DB: ${POSTGRES_DB}
@@ -31,16 +31,16 @@ services:
       - .env
     environment:
       # перекрывает appsettings.Development.json:
-      ConnectionStrings__Default: ${ConnectionStrings__Default}
+      ConnectionStrings__Default: "Host=db;Port=${POSTGRES_PORT_INT};Database=${POSTGRES_DB};Username=${APP_USER};Password=${APP_PASSWORD};Pooling=true"
       ASPNETCORE_ENVIRONMENT: ${ASPNETCORE_ENVIRONMENT}
-      ASPNETCORE_URLS: ${ASPNETCORE_URLS}
+      ASPNETCORE_URLS: "http://+:${HTTP_PORT_INT}"
     depends_on:
       db:
         condition: service_healthy
     ports:
-      - "${HTTP_PORT}:8080"
+      - "${HTTP_PORT_EXT}:${HTTP_PORT_INT}"
     healthcheck:
-      test: ["CMD-SHELL", "curl -fsS http://localhost:8080/health || exit 1"]
+      test: ["CMD-SHELL", "curl -fsS http://localhost:${HTTP_PORT_INT}/health || exit 1"]
       interval: 10s
       timeout: 3s
       retries: 12
