+++ src/backend/Api/Properties/launchSettings.json
@@ -11,10 +11,17 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "ApiProd": {
+    "Api-Verbose": {
       "commandName": "Project",
+      "dotnetRunMessages": true,
+      "launchBrowser": true,
+      "launchUrl": "swagger",
       "applicationUrl": "http://localhost:5000",
-      "environmentVariables": { "ASPNETCORE_ENVIRONMENT": "Production" }
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "Logging__LogLevel__Default": "Debug",
+        "Logging__LogLevel__Microsoft.AspNetCore": "Information"
+      }
     }
   }
 }
