+++ src/backend/Api/appsettings.Development.json.example
@@ -7,6 +7,5 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
-  },
-  "UseInMemoryRepository": false
+  }
 }
