+++ src/backend/Api/appsettings.json
@@ -1,13 +1,9 @@
 {
-  "ConnectionStrings": {
-    "Default": "Host=db;Port=5432;Database=workshopcode;Username=wc_app;Password=wc_app_pwd;"
-  },
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+      "Default": "Warning",
+      "Microsoft.AspNetCore": "Error"
     }
   },
-  "AllowedHosts": "*",
-  "UseInMemoryRepository": false
+  "AllowedHosts": "*"
 }
