+++ src/backend/Infrastructure.Postgres/Identity/.gitignore
@@ -1,0 +1,8 @@
+﻿# Оставить папку в репозитории, но игнорировать её содержимое (только заглушка)
+#*
+#!.gitignore
+
+# В будущем здесь будет храниться слой EF Core для DbContext Indentity
+#   IndentityDbContext.cs
+#   Repositories/				— репозитории для сущностей DbContext Indentity
+#   Migrations/                 — миграции для DbContext Indentity
