+++ src/backend/Infrastructure.Postgres/Infrastructure.Postgres.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-	<RootNamespace>Infrastructure</RootNamespace>
+	<RootNamespace>Infrastructure.Postgres</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
