+++ src/backend/Infrastructure.Postgres/InfrastructureConfigurations.cs
@@ -1,8 +1,6 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 
-namespace Api;
+namespace Infrastructure.Postgres;
 
 public static class InfrastructureConfigurations
 {
