+++ src/backend/Infrastructure.Postgres/Time/TimeDbContextFactory.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Design;
+using Microsoft.Extensions.Configuration;
+
+namespace Infrastructure.Postgres.Time;
+
+public sealed class TimeDbContextFactory : IDesignTimeDbContextFactory<TimeDbContext>
+{
+    public TimeDbContext CreateDbContext(string[] args)
+    {
+        var cfg = InfrastructureConfigurations.GetConfiguration();
+
+        var cs = cfg.GetConnectionString("Default")
+                 ?? Environment.GetEnvironmentVariable("ConnectionStrings__Default");
+
+        Console.WriteLine($"[TimeDbContextFactory]: ConnectionStrings=[{cs}]");
+        
+            var opts = new DbContextOptionsBuilder<TimeDbContext>()
+            .UseNpgsql(cs, x => x.EnableRetryOnFailure())
+            .Options;
+
+        return new TimeDbContext(opts);
+    }
+}
