--- src/backend/Infrastructure/Mono/.gitignore
@@ -1,12 +1,0 @@
-﻿# Оставить папку в репозитории, но игнорировать её содержимое (только заглушка)
-*
-!.gitignore
-
-# В будущем здесь будет храниться слой для MongoDB (NoSQL), напр. чат/вложения:
-# Mongo/
-#   MongoClientFactory.cs
-#   Collections/
-#     ExerciseMessages.cs
-#     Attachments.cs
-#   Storages/
-#     MongoFileStorage.cs      — реализует IFileStorage
