--- src/backend/Infrastructure/Postgres/.gitignore
@@ -1,12 +1,0 @@
-﻿# Оставить папку в репозитории, но игнорировать её содержимое (только заглушка)
-#*
-#!.gitignore
-
-# В будущем здесь будет храниться слой EF Core для PostgreSQL:
-# Postgres/                      — реляционная персистентность (EF Core)
-#   AppDbContext.cs
-#   Repositories/
-#     CourseRepository.cs        — реализует ICourseRepository
-#     StudentExerciseRepository.cs
-#   UnitOfWork.cs                — реализует IUnitOfWork (DbContext.SaveChanges)
-#   Migrations/                  — миграции EF Core
